SSR/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
djangorestframework==3.12.4
drf-spectacular==0.17.3
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
zope.interface==5.4.0
